mejora de vistas

+++ front/src/app/product/[productId]/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import { useState, useEffect } from "react";
 import Swal from "sweetalert2";
 import { IoHome } from "react-icons/io5";
-import Image from "next/image";
+import { FaCartArrowDown } from "react-icons/fa";
 
 const DetailsProduct = ({ params }: { params: { productId: string } }) => {
   const router = useRouter();
@@ -23,7 +23,7 @@ const DetailsProduct = ({ params }: { params: { productId: string } }) => {
     fetchProduct();
   }, [params.productId]);
 
-  //!Agregar al carrito
+  //! Agregar al carrito
   const handleBuyClickAgregar = () => {
     const userToken = localStorage.getItem("userSession");
     if (!userToken) {
@@ -41,13 +41,23 @@ const DetailsProduct = ({ params }: { params: { productId: string } }) => {
       });
     } else {
       const currentCart = JSON.parse(localStorage.getItem("cart") || "[]");
-      currentCart.push({ ...product, quantity });
-      localStorage.setItem("cart", JSON.stringify(currentCart));
-      Swal.fire({
-        icon: "info",
-
-        text: "Producto agregado al carrito.",
-      });
+      const existingProduct = currentCart.find(
+        (item: any) => item.id === params.productId
+      );
+      if (existingProduct) {
+        Swal.fire({
+          icon: "info",
+          text: "El producto ya existe en el carrito.",
+        });
+        router.push("/cart"); // Redirige a la página del carrito
+      } else {
+        currentCart.push({ ...product, quantity });
+        localStorage.setItem("cart", JSON.stringify(currentCart));
+        Swal.fire({
+          icon: "info",
+          text: "Producto agregado al carrito.",
+        });
+      }
     }
   };
   //!Pagar
@@ -84,78 +94,76 @@ const DetailsProduct = ({ params }: { params: { productId: string } }) => {
   }
 
   return (
-    <div className="max-w-4xl mx-auto p-8 md:h-screen">
-      <div className="flex flex-col justify-between lg:flex-row gap-16 lg:items-center">
-        <div className="flex flex-col gap-6 lg:w-2/4">
-          <img
-            src={product.image}
-            alt=""
-            className="w-72 h-96 aspect-square object-cover rounded-xl"
-          />
-        </div>
-
-        <div className="flex flex-col gap-4 lg:w-2/4">
-          <div>
-            <span className="text-violet-600 font-semibold">
-              Ofertas SmartMarket
-            </span>
-            <h1 className="text-3xl font-bold">{product.name}</h1>
-          </div>
-          <p className="text-gray-700">{product.description}</p>
-          <h4 className="text-2xl font-semibold">Stock: {product.stock}</h4>
-          <div>
-            <h1 className="text-3xl text-green-500 font-bold">
-              Precio: $ {product.price}
-            </h1>
-          </div>
-          <div className="flex flex-row items-center gap-12">
-            <div className="flex flex-row items-center">
-              <button
-                className="bg-gray-200 py-2 px-5 rounded-lg text-violet-800 text-3xl"
-                onClick={() => setQuantity(quantity - 1)}
-                disabled={quantity <= 1}
-              >
-                -
-              </button>
-              <span className="py-4 px-6 rounded-lg">{quantity}</span>
-              <button
-                className="bg-gray-200 py-2 px-4 rounded-lg text-violet-800 text-3xl"
-                onClick={() => setQuantity(quantity + 1)}
-                disabled={quantity >= product.stock}
-              >
-                +
-              </button>
-            </div>
-            <button
-              className="flex gap-2 bg-cyan-800 hover:bg-cyan-700 text-white font-semibold py-3 px-8 rounded-xl h-full"
-              onClick={handleBuyClickAgregar}
-            >
-              Agregar
-              <Image
-                src="/cart.png"
-                className="w-6 h-6"
-                alt="cart"
-                width={30}
-                height={25}
-              />
-            </button>
-          </div>
-          <div className="flex flex-row items-center gap-12">
-            <button
-              className="bg-yellow-800 hover:bg-yellow-500 text-white font-semibold py-3 px-6 rounded-xl h-full"
-              onClick={handleBuyClickPagar}>
-              Ir a pagar
-            </button>
-            <a
-              className=" text-sky-700 flex flex-row items-center underline"
-              href="/product">
-              Ir a la tienda
-              <IoHome/>
-            </a>
+   
+    <div className="max-w-4xl mx-auto p-8 md:pt-6 md:h-screen">
+    <div className="flex flex-col md:flex-row gap-11 py-10 px-5 bg-teal-50 rounded-md shadow-lg w-full md:max-w-2xl">
+      <div className="text-sky-600 flex flex-col justify-between">
+        <img
+          src={product.image}
+          alt={product.name}
+          className="w-full h-auto md:w-96 md:h-auto rounded-lg "
+        />
+        <div>
+          <small className="uppercase">Disponibles: </small>
+          <div className="flex flex-wrap md:flex-nowrap gap-1">
+            {product.stock} und.
           </div>
         </div>
       </div>
+      <div className="text-sky-600">
+        <small className="uppercase">Oferta SmartMarket</small>
+        <h3 className="uppercase text-black text-2xl font-medium">
+          {product.name}
+        </h3>
+        <h3 className="text-2xl font-semibold mb-7">${product.price}</h3>
+        <div className="flex items-center">
+          <button
+            className="bg-gray-200 py-0.5 px-2 rounded-lg text-sky-600 text-3xl"
+            onClick={() => setQuantity(quantity - 1)}
+            disabled={quantity <= 1}
+          >
+            -
+          </button>
+          <span className="px-3 rounded-lg">{quantity}</span>
+          <button
+            className="bg-gray-200 py-0.5 px-2 rounded-lg text-sky-600 text-3xl"
+            onClick={() => setQuantity(quantity + 1)}
+            disabled={quantity >= product.stock}
+          >
+            +
+          </button>
+        </div>
+        <br />
+        <small className="text-black">{product.description}</small>
+  
+        <div className="flex flex-col md:flex-row gap-0.5 mt-4">
+          <button
+            id="addToCartButton"
+            onClick={handleBuyClickAgregar}
+            className="bg-yellow-500 flex flex-row justify-center items-center hover:bg-yellow-400 focus:outline-none transition text-white uppercase px-8 py-3"
+          >
+            Agregar
+            <FaCartArrowDown />
+          </button>
+          <button
+            id="addToCartButton"
+            onClick={handleBuyClickPagar}
+            className="bg-yellow-500 hover:bg-yellow-400 focus:outline-none transition text-white uppercase px-8 py-3"
+          >
+            Comprar ahora
+          </button>
+        </div>
+        <a
+          className="text-sky-600 flex flex-row items-center underline mt-4 md:mt-0"
+          href="/product"
+        >
+          Ir a la tienda
+          <IoHome />
+        </a>
+      </div>
     </div>
+  </div>
+  
   );
 };
